% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTCR.R
\name{AddTCR}
\alias{AddTCR}
\title{Add TCR to Seurat object
This function adds Cell Ranger TCR output information with transcriptome, cell hashing,
and others if applicable
To output a comprehensive table to contain all information that originate from a single
barcode}
\usage{
AddTCR(clono, consensus, filtered_barcodes)
}
\arguments{
\item{clono}{clonotypes.csv from Cell Ranger}

\item{consensus}{consensus_annotation.csv from Cell Ranger}

\item{filtered_barcodes}{filtered_contig_annotations.csv from Cell Ranger}
}
\value{
a meta table contains all TCR info
Check how they solve duplicated barcodes
Using the samples and ID parameters, the
function will add the strings as prefixes to prevent issues with repeated
barcodes
They used samples and ID to label barcodes in order to distinguish repeated barcodes
among different samples/IDs.
However, in my previous cases, duplicated barcodes were observed within one single
run/library
}
\description{
Three CSV files generated from Cell Ranger are required as input files for this function
}
