% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTCR_details.R
\name{AddTCR_details}
\alias{AddTCR_details}
\title{Consolidates Cell Ranger VDJ outputs (>= cellranger-6.0.0)}
\usage{
AddTCR_details(clono, consensus, filtered_barcodes)
}
\arguments{
\item{clono}{clonotypes.csv from Cell Ranger}

\item{consensus}{consensus_annotation.csv from Cell Ranger}

\item{filtered_barcodes}{filtered_contig_annotations.csv from Cell Ranger}
}
\value{
a meta table contains all TCR info
}
\description{
This function adds Cell Ranger TCR output information with transcriptome, cell hashing,
and others if applicable
Also factor in different patterns of clonotype pairings
To output a comprehensive table to contain all information that originate from a single
barcode
}
\details{
Three CSV files generated from Cell Ranger are required as input files for this function
}
